# server/docs/swagger.yaml
definitions:
...
  schemas.CircuitShort:
    properties:
      RxBandWidth:
        type: integer
      TxBandWidth:
        type: integer
      id:
        type: string
      name:
        type: string
      provider:
        type: string
    type: object
...
  schemas.IdResponse:
    properties:
      id:
        type: string
    type: object
  schemas.ListResponse:
    properties:
      results: {}
      total:
        type: integer
    type: object
...
  schemas.Prefix:
    properties:
      createdAt:
        type: string
      id:
        type: string
      range:
        type: string
      siteId:
        type: string
      type:
        type: string
      updatedAt:
        type: string
      utilization:
        type: number
      version:
        type: string
      vlanId:
        type: integer
      vlanName:
        type: string
    type: object
  schemas.PrefixCreate:
    properties:
      range:
        type: string
      siteId:
        type: string
      type:
        enum:
        - Dynamic
        - Static
        type: string
      vlanId:
        description: not support VxLAN now
        maximum: 4094
        type: integer
      vlanName:
        type: string
    required:
    - range
    - siteId
    - type
    type: object
  schemas.PrefixUpdate:
    properties:
      range:
        type: string
      siteId:
        type: string
      type:
        enum:
        - Dynamic
        - Static
        type: string
      vlanId:
        maximum: 4094
        type: integer
      vlanName:
        type: string
    type: object
...
  schemas.Site:
    properties:
      address:
        type: string
      createdAt:
        type: string
      description:
        type: string
      id:
        type: string
      latitude:
        type: number
      longitude:
        type: number
      name:
        type: string
      region:
        type: string
      siteCode:
        type: string
      status:
        type: string
      timeZone:
        type: string
      updatedAt:
        type: string
    type: object
  schemas.SiteCreate:
    properties:
      address:
        type: string
      description:
        type: string
      latitude:
        type: number
      longitude:
        type: number
      name:
        type: string
      region:
        type: string
      siteCode:
        type: string
      status:
        enum:
        - Active
        - Inactive
        type: string
      timeZone:
        type: string
    required:
    - address
    - latitude
    - longitude
    - name
    - region
    - siteCode
    - status
    - timeZone
    type: object
  schemas.SiteDetail:
    properties:
      address:
        type: string
      apCount:
        type: integer
      circuit:
        items:
          $ref: '#/definitions/schemas.CircuitShort'
        type: array
      circuitCount:
        type: integer
      createdAt:
        type: string
      description:
        type: string
      gatewayCount:
        type: integer
      id:
        type: string
      latitude:
        type: number
      longitude:
        type: number
      name:
        type: string
      rackCount:
        type: integer
      region:
        type: string
      siteCode:
        type: string
      status:
        type: string
      switchCount:
        type: integer
      timeZone:
        type: string
      updatedAt:
        type: string
      vlanCount:
        type: integer
    type: object
  schemas.SiteUpdate:
    properties:
      address:
        type: string
      description:
        type: string
      latitude:
        type: number
      longitude:
        type: number
      name:
        type: string
      region:
        type: string
      siteCode:
        type: string
      timeZone:
        type: string
    type: object
...
info:
  contact:
    name: Jeffry
  description: Narvis OpenAPI for developers.
  termsOfService: http://swagger.io/terms/
  title: Narvis API
  version: "1.0"
paths:
...
  /infra/sites:
    get:
      consumes:
      - application/json
      description: List sites
      parameters:
      - collectionFormat: csv
        in: query
        items:
          type: string
        name: id
        type: array
      - in: query
        name: keyword
        type: string
      - collectionFormat: csv
        in: query
        items:
          type: string
        name: name
        type: array
      - enum:
        - asc
        - desc
        in: query
        name: order
        type: string
      - in: query
        name: orderBy
        type: string
      - in: query
        minimum: 1
        name: page
        type: integer
      - in: query
        maximum: 1000
        minimum: 1
        name: pageSize
        type: integer
      - collectionFormat: csv
        in: query
        items:
          type: string
        name: region
        type: array
      - collectionFormat: csv
        in: query
        items:
          type: string
        name: siteCode
        type: array
      - enum:
        - Active
        - Inactive
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/schemas.ListResponse'
            - properties:
                results:
                  items:
                    $ref: '#/definitions/schemas.Site'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: List sites
      tags:
      - Infra
    post:
      consumes:
      - application/json
      description: Create new site
      parameters:
      - description: site
        in: body
        name: site
        required: true
        schema:
          $ref: '#/definitions/schemas.SiteCreate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.IdResponse'
      security:
      - BearerAuth: []
      summary: Create new site
      tags:
      - Infra
  /infra/sites/{id}:
    delete:
      consumes:
      - application/json
      description: Delete site
      parameters:
      - description: site id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.IdResponse'
      security:
      - BearerAuth: []
      summary: Delete site
      tags:
      - Infra
    get:
      consumes:
      - application/json
      description: Get site
      parameters:
      - description: site id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.SiteDetail'
      security:
      - BearerAuth: []
      summary: Get site
      tags:
      - Infra
    put:
      consumes:
      - application/json
      description: Update site
      parameters:
      - description: site id
        in: path
        name: id
        required: true
        type: string
      - description: site
        in: body
        name: site
        required: true
        schema:
          $ref: '#/definitions/schemas.SiteUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.IdResponse'
      security:
      - BearerAuth: []
      summary: Update site
      tags:
      - Infra
...
  /ipam/prefixes:
    get:
      consumes:
      - application/json
      description: Get prefix list
      parameters:
      - in: query
        name: keyword
        type: string
      - enum:
        - asc
        - desc
        in: query
        name: order
        type: string
      - in: query
        name: orderBy
        type: string
      - in: query
        minimum: 1
        name: page
        type: integer
      - in: query
        maximum: 1000
        minimum: 1
        name: pageSize
        type: integer
      - collectionFormat: csv
        in: query
        items:
          type: string
        name: range
        type: array
      - in: query
        name: siteId
        type: string
      - enum:
        - Dynamic
        - Static
        in: query
        name: type
        type: string
      - collectionFormat: csv
        in: query
        items:
          type: integer
        maxItems: 4094
        name: vlanId
        type: array
      - collectionFormat: csv
        in: query
        items:
          type: string
        name: vlanName
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/schemas.ListResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/schemas.Prefix'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: Get prefix list
      tags:
      - IPAM
    post:
      consumes:
      - application/json
      description: Create prefix
      parameters:
      - description: Prefix
        in: body
        name: prefix
        required: true
        schema:
          $ref: '#/definitions/schemas.PrefixCreate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.IdResponse'
      security:
      - BearerAuth: []
      summary: Create prefix
      tags:
      - IPAM
  /ipam/prefixes/{id}:
    delete:
      consumes:
      - application/json
      description: Delete prefix
      parameters:
      - description: prefixId
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.IdResponse'
      security:
      - BearerAuth: []
      summary: Delete prefix
      tags:
      - IPAM
    get:
      consumes:
      - application/json
      description: Get prefix
      parameters:
      - description: prefixId
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.Prefix'
      security:
      - BearerAuth: []
      summary: Get prefix
      tags:
      - IPAM
    put:
      consumes:
      - application/json
      description: Update prefix
      parameters:
      - description: prefixId
        in: path
        name: id
        required: true
        type: string
      - description: Prefix
        in: body
        name: prefix
        required: true
        schema:
          $ref: '#/definitions/schemas.PrefixUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.IdResponse'
      security:
      - BearerAuth: []
      summary: Update prefix
      tags:
      - IPAM
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"